Exercicio 4 do cap 03

+++ python/cap_03/exercicio_3_4.py
@@ -1,4 +1,4 @@
-salario = int(input("Digite seu salário: "))
+salario = float(input("Digite seu salário: "))
 if (salario >= 1200):
     print("Paga IR")
 else:
